requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asyncmy==0.2.9
bcrypt==4.2.0
black==24.8.0
click==8.1.7
databases==0.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.114.0
h11==0.14.0
idna==3.8
isort==5.13.2
Mako==1.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.2
pydantic==2.9.0
pydantic_core==2.23.2
PyJWT==2.9.0
PyMySQL==1.1.1
python-multipart==0.0.9
pytz==2024.1
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
